% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{frsr_bin}
\alias{frsr_bin}
\title{FRSR Bin}
\usage{
frsr_bin(
  x_min = 0.25,
  x_max = 1,
  n_bins = 4,
  NRmax = 0,
  float_samples = 1024,
  magic_samples = 2048,
  magic_min = 1596980000L,
  magic_max = 1598050000L
)
}
\arguments{
\item{x_min}{Numeric. Default is 0.25.}

\item{x_max}{Numeric. Default is 1.0.}

\item{n_bins}{Integer. The number of bins to divide the range into. Default is 4.}

\item{NRmax}{Integer. The maximum number of Newton-Raphson iterations (default: 0).}

\item{float_samples}{Integer. The number of floating-point samples generated per bin.}

\item{magic_samples}{Integer. The number of magic constant samples to generate per bin.}

\item{magic_min}{Integer. The minimum magic constant to test (default: 1596980000).}

\item{magic_max}{Integer. The maximum magic constant to test (default: 1598050000).}
}
\value{
A data frame with columns:
    \item{N_bins}{Total number of bins}
    \item{Location}{Bin number (1-indexed)}
    \item{Range_Min}{Minimum value of the bin range}
    \item{Range_Max}{Maximum value of the bin range}
    \item{Magic}{Optimal magic constant as an integer}
    \item{Max_Relative_Error}{Maximum relative error for each bin}
}
\description{
Generate optimal magic constants for the Fast Reciprocal Square Root algorithm over specified bins.
}
\details{
This function divides the range [x_min, x_max] into n_bins bins and generates float_samples
floating-point samples within each bin. It then tests magic_samples magic constants within
the range [magic_min, magic_max] to find the optimal magic constant for each bin that minimizes
the sum relative error per bin. The achievable error and optimal constants will change with
bin size and number of bins, as well as the integer and float samples. 

The FRSR is periodic over 0.25 to 1.0 with a magic constant of 0x5f3759df. I don't know if 
this is generally true.
}
\examples{
\donttest{
# Generate optimal magic constants for the range [0.25, 1.0] divided into 4 bins
result <- frsr_bin()
print(result)
#   Location Range_Min Range_Max      Magic    Sum_Error N_bins
# 1        1 0.2500000 0.3535534 1597413411 4.549199e-04      4
# 2        2 0.3535534 0.5000000 1597115686 6.175506e-05      4
# 3        3 0.5000000 0.7071068 1597150657 3.486200e-05      4
# 4        4 0.7071068 1.0000000 1597488127 8.389181e-04      4
# }
}
